Improved a bit logic for wordpress event trigger
