+++ internal/dictionary/slack_dictionary.json
@@ -1,1 +1,1 @@
-{"text_message_dictionary":[{"question":"(?i)(?:release (\\w+))","answer":"Ok. I will release %s branch","main_group_index_in_regex":1,"reaction_type":"release_pipeline_trigger"},{"question":"(?i)(?:release ([a-z-]+))","answer":"Okay, soon %s branch will be deployed","main_group_index_in_regex":1,"reaction_type":"release_pipeline_trigger"},{"question":"(?:release `([a-z-]+)`)","answer":"Ok sir, I will release %s branch","main_group_index_in_regex":1,"reaction_type":""},{"question":"(?i)Hey","answer":"Yo","main_group_index_in_regex":0,"reaction_type":""},{"question":"(?i)Hello","answer":"Yo","main_group_index_in_regex":0,"reaction_type":""},{"question":"(?i)Welcome","answer":"Thank you","main_group_index_in_regex":0,"reaction_type":""},{"question":"(?i)hi","answer":"Hello hello","main_group_index_in_regex":0,"reaction_type":""},{"question":"(?i)what you can","answer":"So, here is the list of possible functions available for now: \n - release pipeline triggering","main_group_index_in_regex":0,"reaction_type":"help"},{"question":"(?i)I have question","answer":"Whatsup?","main_group_index_in_regex":0,"reaction_type":""}],"file_message_dictionary":[{"question":"(?i)process","answer":"I received the file, so now I will process it and will send the result back. Please wait a bit","main_group_index_in_regex":0,"reaction_type":"themer_event"}]}
+{"text_message_dictionary":[{"question":"(?i)(?:release (\\w+))","answer":"Ok. I will release %s branch","main_group_index_in_regex":1,"reaction_type":"release_pipeline_trigger"},{"question":"(?i)(?:release ([a-z-]+))","answer":"Okay, soon %s branch will be deployed","main_group_index_in_regex":1,"reaction_type":"release_pipeline_trigger"},{"question":"(?:release `([a-z-]+)`)","answer":"Ok sir, I will release %s branch","main_group_index_in_regex":1,"reaction_type":""},{"question":"(?i)Hey","answer":"Yo","main_group_index_in_regex":0,"reaction_type":""},{"question":"(?i)Hello","answer":"Yo","main_group_index_in_regex":0,"reaction_type":""},{"question":"(?i)Welcome","answer":"Thank you","main_group_index_in_regex":0,"reaction_type":""},{"question":"(?i)hi","answer":"Hello hello","main_group_index_in_regex":0,"reaction_type":""},{"question":"(?i)what you can","answer":"So, here is the list of possible functions available for now: \n - release pipeline triggering","main_group_index_in_regex":0,"reaction_type":"help"},{"question":"(?i)I have question","answer":"Whatsup?","main_group_index_in_regex":0,"reaction_type":""}],"file_message_dictionary":[{"question":"(?i)process","answer":"I received the file, so now I will process it and will send the result back. Please wait a bit","main_group_index_in_regex":0,"reaction_type":"themer_wordpress_event"}]}
